remove accessibility event throttling in Progressbar
The current throttling is not good: when we update progress periodically
at an interval smaller than 200ms, no progress is announced except
for the last 100% in talkback. We will reply on the throttling when
sending out state change events after adopting the accessibility state
api: we only send out state change events when the state description
actually changes.

Bug: 143079143
Test: tested with talkback test when updating progress at different
frequencies.

Change-Id: I7480a3420130fc42645df17074b093cdced430ae