Add PacProcessor system API.
PacProcessor is intended to be used as an internal implementation detail of
com.android.pacprocessor.PacService, which in turn implements the java.net.ProxySelector interface.

Processing the PAC file requires starting up an instance of V8
so is expected to be handled by the central system service and queried over IPC (PacService),
not for apps to try to run it themselves.

Bug: 147578322
Test: N/A.

Change-Id: I857bb3616d7029fa0d42bf25def50f46094ec1d9